+++ src/components/ServiceFormEditor/FieldEditor/ChangeServiceTypeDropdown.tsx
@@ -1,5 +1,5 @@
 import { Select } from '@radix-ui/themes';
-import { FormFieldType, formFieldTypes } from './formField';
+import { FormFieldType, formFieldTypes } from '../../../classes/service/formField';
 import ServiceType from './ServiceType';
 
 export default function ChangeServiceTypeDropdown(props: {
