+++ src/components/ServiceFormEditor/FieldEditor/CheckboxFieldEditor.tsx
@@ -1,18 +1,17 @@
-import { Flex, TextField, Button } from '@radix-ui/themes';
-import { RadioField, FormFieldType } from './formField';
-import React from 'react';
+import { Button, Flex, TextField } from '@radix-ui/themes';
+import { CheckboxField, FormFieldType } from '../../../classes/service/formField';
 import NoOutlineIconButton from './NoOutlineIconButton';
 import { Cross1Icon, PlusIcon } from '@radix-ui/react-icons';
 
-export default function RadioFormField(props: {
-  field: RadioField;
+export default function CheckboxFieldEditor(props: {
+  field: CheckboxField;
   onChange: (field: FormFieldType) => void | undefined;
 }) {
   function editChoice(index: number) {
     return (evt: React.ChangeEvent<HTMLInputElement>) => {
       const newChoices = [...props.field.choices];
       newChoices.splice(index, 1, evt.currentTarget.value);
-      const newField = new RadioField(props.field.prompt, newChoices);
+      const newField = new CheckboxField(props.field.prompt, newChoices);
       props.onChange(newField);
     };
   }
@@ -23,7 +22,7 @@ export default function RadioFormField(props: {
       `Choice ${props.field.choices.length + 1}`,
     ];
 
-    const newField = new RadioField(props.field.prompt, newChoices);
+    const newField = new CheckboxField(props.field.prompt, newChoices);
     props.onChange(newField);
   }
 
@@ -35,7 +34,7 @@ export default function RadioFormField(props: {
       }
       const newChoices = [...props.field.choices];
       newChoices.splice(index, 1);
-      const newField = new RadioField(props.field.prompt, newChoices);
+      const newField = new CheckboxField(props.field.prompt, newChoices);
       props.onChange(newField);
     };
   }
