+++ src/components/ServiceFormEditor/FieldEditor/RadioFieldEditor.tsx
@@ -1,17 +1,18 @@
-import { Button, Flex, TextField } from '@radix-ui/themes';
-import { CheckboxField, FormFieldType } from './formField';
+import { Flex, TextField, Button } from '@radix-ui/themes';
+import { RadioField, FormFieldType } from '../../../classes/service/formField';
+import React from 'react';
 import NoOutlineIconButton from './NoOutlineIconButton';
 import { Cross1Icon, PlusIcon } from '@radix-ui/react-icons';
 
-export default function CheckboxFieldEditor(props: {
-  field: CheckboxField;
+export default function RadioFormField(props: {
+  field: RadioField;
   onChange: (field: FormFieldType) => void | undefined;
 }) {
   function editChoice(index: number) {
     return (evt: React.ChangeEvent<HTMLInputElement>) => {
       const newChoices = [...props.field.choices];
       newChoices.splice(index, 1, evt.currentTarget.value);
-      const newField = new CheckboxField(props.field.prompt, newChoices);
+      const newField = new RadioField(props.field.prompt, newChoices);
       props.onChange(newField);
     };
   }
@@ -22,7 +23,7 @@ export default function CheckboxFieldEditor(props: {
       `Choice ${props.field.choices.length + 1}`,
     ];
 
-    const newField = new CheckboxField(props.field.prompt, newChoices);
+    const newField = new RadioField(props.field.prompt, newChoices);
     props.onChange(newField);
   }
 
@@ -34,7 +35,7 @@ export default function CheckboxFieldEditor(props: {
       }
       const newChoices = [...props.field.choices];
       newChoices.splice(index, 1);
-      const newField = new CheckboxField(props.field.prompt, newChoices);
+      const newField = new RadioField(props.field.prompt, newChoices);
       props.onChange(newField);
     };
   }
